+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- New `antora` template for generating AsciiDoc API documentation for use in Antora.
+
 ## 0.1.0 - 2021-10-12
 
 ### Improved